refactor : infrastructure 패키지 도메인별 그룹핑 및 Adapter 네이밍 통일
