+++ oop-infrastructure/src/main/kotlin/com/jaeyong/oop/infrastructure/health/repository/HealthEntityRepository.kt
@@ -1,0 +1,7 @@
+package com.jaeyong.oop.infrastructure.health.repository
+
+import com.jaeyong.oop.infrastructure.health.entity.HealthEntity
+
+interface HealthEntityRepository {
+    fun save(entity: HealthEntity): HealthEntity
+}
