+++ oop-infrastructure/src/main/kotlin/com/jaeyong/oop/infrastructure/health/repository/HealthJpaRepository.kt
@@ -1,0 +1,6 @@
+package com.jaeyong.oop.infrastructure.health.repository
+
+import com.jaeyong.oop.infrastructure.health.entity.HealthEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface HealthJpaRepository : JpaRepository<HealthEntity, Long>, HealthEntityRepository
