--- oop-infrastructure/src/main/kotlin/com/jaeyong/oop/infrastructure/persistence/HealthJpaRepository.kt
@@ -1,5 +1,0 @@
-package com.jaeyong.oop.infrastructure.persistence
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface HealthJpaRepository : JpaRepository<HealthEntity, Long>
