+++ oop-infrastructure/src/main/kotlin/com/jaeyong/oop/infrastructure/user/adapter/PasswordEncryptorAdapter.kt
@@ -1,11 +1,11 @@
-package com.jaeyong.oop.infrastructure.user
+package com.jaeyong.oop.infrastructure.user.adapter
 
-import com.jaeyong.oop.domain.user.port.PasswordEncryptor
+import com.jaeyong.oop.domain.user.port.PasswordEncryptorPort
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
 import org.springframework.stereotype.Component
 
 @Component
-class BCryptPasswordEncryptor : PasswordEncryptor {
+class PasswordEncryptorAdapter : PasswordEncryptorPort {
     private val encoder = BCryptPasswordEncoder()
 
     override fun encrypt(raw: String): String = encoder.encode(raw)
